added documentation, nested SettingsScreenComponent to SettingsTrackScreen
