+++ app/build.gradle.kts
@@ -97,10 +97,10 @@ dependencies {
     implementation("androidx.compose.ui:ui-graphics")
     implementation("androidx.compose.ui:ui-tooling-preview")
     implementation("androidx.compose.material3:material3:1.3.0")
-    implementation(platform("androidx.compose:compose-bom:2024.09.01"))
-    androidTestImplementation(platform("androidx.compose:compose-bom:2024.09.01"))
+    implementation(platform("androidx.compose:compose-bom:2024.09.02"))
+    androidTestImplementation(platform("androidx.compose:compose-bom:2024.09.02"))
     androidTestImplementation("androidx.compose.ui:ui-test-junit4")
-    debugImplementation("androidx.compose.ui:ui-tooling:1.7.1")
+    debugImplementation("androidx.compose.ui:ui-tooling:1.7.2")
     debugImplementation("androidx.compose.ui:ui-test-manifest")
     implementation("androidx.core:core-splashscreen:1.0.1")
