+++ app/src/main/java/com/savenko/track/presentation/components/ideasCards/TrackScreenFeed.kt
@@ -36,7 +36,8 @@ import org.koin.compose.koinInject
 
 /**
  * It is main screen feed which contains automatically scrollable pager which always starts
- * with card TrackMainFeedCard and ending with NewIdeaFeedCard
+ * with card [BudgetFeedCard] and ending with [NewIdeaFeedCard]
+ *
  * Scroll and state is handled by [TrackScreenFeedViewModel]
  */
 @Composable
